Fix possible loss of data warning in RGBController_ZotacV2GPU.cpp

+++ Controllers/ZotacV2GPUController/RGBController_ZotacV2GPU.cpp
@@ -451,13 +451,13 @@ void RGBController_ZotacV2GPU::DeviceUpdateMode()
     zoneConfig.colorPreset = 0;
     if(zoneConfig.mode == ZOTAC_V2_GPU_MODE_DUET)
     {
-        zoneConfig.colorPreset = ZOTAC_V2_GPU_DUET_PRESETS.size();  // custom
+        zoneConfig.colorPreset = (unsigned int)ZOTAC_V2_GPU_DUET_PRESETS.size();  // custom
         for(size_t i = 0; i < ZOTAC_V2_GPU_DUET_PRESETS.size(); ++i)
         {
             if(zoneConfig.color1 == ZOTAC_V2_GPU_DUET_PRESETS[i].first &&
                 zoneConfig.color2 == ZOTAC_V2_GPU_DUET_PRESETS[i].second)
             {
-                zoneConfig.colorPreset = i;
+                zoneConfig.colorPreset = (unsigned int)i;
             }
         }
     }
